% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milonga.R
\name{milonga}
\alias{milonga}
\title{Get milonga playlist info}
\usage{
milonga(
  hrs = 4,
  tanda = 3,
  cycle = 6,
  song = 2.5,
  cortina = 45,
  cortina_actual = cortina,
  start = 8,
  AM = FALSE,
  cumparsita_separate = FALSE,
  chacarera = 0,
  songs_outside = 0,
  songs_outside_mins = cortina_actual
)
}
\arguments{
\item{hrs}{Duration of milonga in hours.}

\item{tanda}{Number of songs per tanda.}

\item{cycle}{Number of tandas in a tanda-cycle. E.g., a TTVTTM structure means 6 tandas in a cycle.}

\item{song}{Average duration of songs in minutes.}

\item{cortina}{Average duration of cortina in seconds.}

\item{cortina_actual}{Average actual duration of cortina in seconds - in case you will only play part of the cortina song (e.g., 45 seconds of a 3 minutes song.}

\item{start}{Start time of the milonga, as an integer (e.g., 8.5 is 8:30).}

\item{AM}{Whether the milonga starts AM rather than PM}

\item{cumparsita_separate}{Is the Cumparsita an extra song separate from the last tanda?}

\item{chacarera}{Number of chacarera songs (not yet implemented).}

\item{songs_outside}{Number of songs outside of the milonga. For instance, warm-up songs before the start, or party songs after the end.}

\item{songs_outside_mins}{Average duration of songs outside of the milonga, in minutes.}
}
\value{
Number of songs in the playlist.
}
\description{
Get milonga playlist info
}
\examples{
milonga() # sensible defaults
milonga(
  hrs = 5, tanda = 4,
  song = 2.7, cortina = 30, cortina_actual = 3.5,
  start = 11, AM = TRUE
  )
milonga(
  hrs = 3.5, tanda = 3, cycle = 5,
  song = 2.4, cortina = 30, cortina_actual = 3.5,
  start = 11, AM = TRUE,
  cumparsita_separate = TRUE,
  chacarera = 1,
  songs_outside = 8, songs_outside_mins = 3.1
  )
}
